+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0] - 2025-05-15
+
+_Stable release based on [2.0.0-rc.1]._
+
+## [2.0.0-rc.1] - 2025-05-15
+
+### Added
+
+- Use `IgMailable` for sending feedback emails.
+
+### Changed
+
+- Update laravel-common version to `^2`.
+- Make email field optional and use it for replyto.
+
+### Removed
+
+- Remove subject field.
+
 ## [1.0.0] - 2025-05-09
 
 _Stable release based on [1.0.0-rc.1]._
@@ -97,6 +116,8 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[2.0.0]: https://https://github.com/internetguru/laravel-feedback/compare/v1.0.0...v2.0.0
+[2.0.0-rc.1]: https://github.com/internetguru/laravel-feedback/releases/tag/v1.0.0
 [1.0.0]: https://https://github.com/internetguru/laravel-feedback/compare/v0.4.1...v1.0.0
 [1.0.0-rc.1]: https://github.com/internetguru/laravel-feedback/releases/tag/v0.4.1
 [0.4.1]: https://https://github.com/internetguru/laravel-feedback/compare/v0.4.0...v0.4.1